Rewrite the UploadZone component in typescript
