+++ src/components/UploadZone/index.ts
@@ -1,0 +1,2 @@
+export { UploadZone, UploadZone as default } from './UploadZone';
+export type { UploadZoneProps, UploadZoneHandler } from './UploadZone';
