+++ src/components/UploadZone/readme.md
@@ -1,12 +1,11 @@
 ```jsx
 <UploadZone
   handler={console.log}
-  subTitle="Second string"
+  subTitle="Support for a single or bulk upload. \nStrictly prohibit from uploading company data or other band files."
 />
 <br />
 <UploadZone
   handler={console.log}
-  subTitle="Second string"
   loading
 />
 ```
