+++ src/index.ts
@@ -3,4 +3,5 @@ export * from './components/Button';
 export * from './components/PageSection';
 export * from './components/SideMenu';
 export * from './components/Text';
+export * from './components/UploadZone';
 export * from './variables';
